refactor: decompose AuthRequests
